pre-commit

+++ functionfuse/serializers/daskarray.py
@@ -5,9 +5,7 @@ import zarr, h5py
 import posixpath, s3fs
 
 
-
 class DaskArraySerializer:
-    
     datapath = "dask_arrays"
     name = "HDFSerializer"
     serializer_class = Array
@@ -20,16 +18,16 @@ class DaskArraySerializer:
         root = zarr.group(store=store)
         return root
 
-
     @classmethod
-    def pickle(cls, obj : Array, protocols : dict):
-
+    def pickle(cls, obj: Array, protocols: dict):
         if FILE_PROTOCOL in protocols:
             file = protocols[FILE_PROTOCOL]
             filename = generate_name()
             f = file.open(filename, "wb")
-            hdf = h5py.File(f, mode = "w")
-            dset = hdf.create_dataset(cls.datapath, obj.shape, dtype= obj.dtype, chunks = obj.chunksize)
+            hdf = h5py.File(f, mode="w")
+            dset = hdf.create_dataset(
+                cls.datapath, obj.shape, dtype=obj.dtype, chunks=obj.chunksize
+            )
             da.store(obj, dset)
             hdf.close()
             f.close()
@@ -37,19 +35,21 @@ class DaskArraySerializer:
             s3 = protocols[S3_PROTOCOL]
             root = cls.get_s3_root(s3)
             filename = generate_name()
-            dset = root.create_dataset(filename, shape=obj.shape, chunks=obj.chunksize, dtype=obj.dtype)
+            dset = root.create_dataset(
+                filename, shape=obj.shape, chunks=obj.chunksize, dtype=obj.dtype
+            )
             da.store(obj, dset)
 
         version = "0.0.0"
         return cls.name, (version, filename)
-    
+
     @classmethod
     def unpickle(cls, pid, protocols: dict):
         _, filename = pid
         if FILE_PROTOCOL in protocols:
             file = protocols[FILE_PROTOCOL]
             f = file.open(filename, "rb")
-            hdf = h5py.File(f, mode = "r")
+            hdf = h5py.File(f, mode="r")
             dset = hdf[cls.datapath]
         else:
             s3 = protocols[S3_PROTOCOL]
@@ -57,7 +57,7 @@ class DaskArraySerializer:
             dset = root[filename]
         array = da.from_array(dset, chunks=dset.chunks)
         return array
-    
+
     @classmethod
     def remove(cls, pid, protocols: dict):
         _, filename = pid
@@ -68,6 +68,6 @@ class DaskArraySerializer:
             s3 = protocols[S3_PROTOCOL]
             client = s3["client"]
             file = posixpath.join(s3["folder"], cls.datapath, filename)
-            client.rm(file, recursive = True)
+            client.rm(file, recursive=True)
 
         return None
